safari-600.7-branch/LayoutTests/inspector-obsolete/elements/edit-style-attribute-expected.txt

Tests that style modification generates attribute updated event only when attribute is actually changed.


Running: testSetUp

Running: testSetNewValue
WebInspector.DOMAgent.Events.AttrModified should be issued

Running: testSetSameValue
WebInspector.DOMNode.prototype._setAttributesPayload should be called
